+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Improve Anthropic API to support more models like kimi-k2.5.
+
 ## 0.97.5
 
 - Support moonshot models via anthropic API.